+++ src/extensions/filter-control/bootstrap-table-filter-control.js
@@ -239,7 +239,7 @@ const UtilsFilterControl = {
             formattedValue = Utils.calculateObjectValue(that.header, column.filterDataCollector, [fieldValue, data[i], formattedValue], formattedValue)
           }
 
-          if (typeof formattedValue === 'object') {
+          if (typeof formattedValue === 'object' && formattedValue !== null) {
             formattedValue.forEach((value) => {
               UtilsFilterControl.addOptionToSelectControl(selectControl, value, value, column.filterDefault)
             })
@@ -766,6 +766,11 @@ $.BootstrapTable = class extends $.BootstrapTable {
     this.EnableControls(false)
     this.onSearch(event, false)
     this.trigger('column-search', $field, text)
+  }
+
+  initToolbar () {
+    this.showSearchClearButton = this.options.filterControl && this.options.showSearchClearButton
+    super.initToolbar()
   }
 
   resetSearch () {