+++ src/themes/materialize/bootstrap-table-materialize.js
@@ -19,7 +19,9 @@ $.extend($.fn.bootstrapTable.defaults, {
     detailOpen: 'add',
     detailClose: 'remove',
     fullscreen: 'fullscreen',
-    sort: 'sort'
+    sort: 'sort',
+    search: 'search',
+    clearSearch: 'delete'
   }
 })
 
@@ -45,6 +47,7 @@ $.BootstrapTable = class extends $.BootstrapTable {
     this.constants.html.pagination = ['<ul class="pagination%s">', '</ul>']
     this.constants.html.paginationItem = '<li class="waves-effect page-item%s" aria-label="%s"><a href="#">%s</a></li>'
     this.constants.html.icon = '<i class="%s">%s</i>'
+    this.constants.html.inputGroup = '%s%s'
   }
 
   initToolbar () {